Move the select device dialog logit to callbacks.c
As we intend to be able to select another device anytime,
via a callback, move the logic associated with device selection
to the callbacks.c.

As a side effect, it makes the main.c cleaner.
